Added abstract server class and client server model
